Use Markdown content and Vocs built-in dark mode
Move portfolio text into index.mdx with standard Markdown syntax.
Remove the custom color scheme hook and rely on Vocs theming, using
CSS dark variants for project card images and letting GitHub calendar
auto-detect the active scheme.
