+++ src/pages/_components/github-contributions.tsx
@@ -1,0 +1,9 @@
+'use client'
+
+import GitHubCalendar from 'react-github-calendar'
+
+export function GitHubContributions() {
+  return (
+    <GitHubCalendar showWeekdayLabels username="mehmetraufoguz" />
+  )
+}
