+++ src/pages/_root.css
@@ -1,0 +1,7 @@
+img[alt='Rauf'] {
+  width: 5rem;
+  height: 5rem;
+  margin-top: 2.5rem;
+  margin-bottom: 2.5rem;
+  border-radius: 9999px;
+}
