+++ src/pages/index.mdx
@@ -1,0 +1,31 @@
+import { FeaturedProjects } from './_components/featured-projects'
+import { GitHubContributions } from './_components/github-contributions'
+
+![Rauf](https://avatars.githubusercontent.com/u/88249309)
+
+# Hello, this is Rauf 👋🏻
+
+> Obi-Wan: Hello there.
+>
+> General Grievous: General Kenobi..
+
+## Contributions
+
+mostly private projects
+
+<GitHubContributions />
+
+## Featured Projects
+
+ta-daa 🥳
+
+<FeaturedProjects />
+
+## Social Links
+
+i dont usually use them
+
+- [Github](https://github.com/mehmetraufoguz)
+- [Linkedin](https://www.linkedin.com/in/mehmetraufoguz)
+- [Telegram](https://t.me/mehmetraufoguz)
+- [Instagram](https://www.instagram.com/mehmetraufoguz)
