--- src/pages/index.tsx
@@ -1,5 +1,0 @@
-import { Portfolio } from './_components/portfolio'
-
-export default function Page() {
-  return <Portfolio />
-}
